More peer id to protocol
